Move latest activity and recent edits views to homepage.views.activity
