+++ homepage/templates/homepage/activity/latest_activity.html
