+++ homepage/tests/test_activity_views.py
